Implemented some TODO on back end

+++ Backend/src/main/java/com/github/trionfettinicoUNICAM/PPTeam_DOIT/service/SimpleProjectsManager.java
@@ -17,8 +17,6 @@ import java.util.*;
 
 @Service
 public class SimpleProjectsManager implements ProjectsManager{
-    //TODO migliorare i controlli e le condizioni sui metodi
-
 
     @Autowired
     private ProjectRepository projectRepository;
@@ -36,14 +34,17 @@ public class SimpleProjectsManager implements ProjectsManager{
 
     @Override
     public Project create(Project project) throws EntityNotFoundException {
-        // TODO: 11/02/2021 estrarre questi controlli per farli anche su update
+        checkProject(project);
+        return projectRepository.save(project);
+    }
+
+    private void checkProject(Project project) throws EntityNotFoundException{
         Organization organization = organizationRepository.findById(project.getOrganizationId()).orElseThrow(()->
                 new EntityNotFoundException("L'organizzazione con ID '"+project.getOrganizationId()+"' " +
                         "passata nel progetto con ID '"+project.getId()+"' non esiste")
         );
         if (!organization.getMembersMails().contains(project.getCreatorMail()))
             throw new IllegalArgumentException("Il creatore del progetto non è tra i membri dell'organizzazione del progetto");
-        return projectRepository.save(project);
     }
 
     @Override
@@ -58,16 +59,17 @@ public class SimpleProjectsManager implements ProjectsManager{
     }
 
     @Override
-    public boolean delete(String projectID) {
+    public boolean delete(String projectID){
         if(projectID.isBlank()) throw new IllegalArgumentException("Il campo 'ID' è vuoto");
-        // TODO: 04/02/2021 se il progetto non esiste lancio illegal argument
+        if (!exists(projectID)) throw new IllegalArgumentException("projectID non presente");
         projectRepository.deleteById(projectID);
         return !exists(projectID);
     }
 
     @Override
-    public Project update(Project project) {
-        // TODO: 04/02/2021 se il progetto non esiste lancio illegal argument
+    public Project update(Project project) throws EntityNotFoundException {
+        if (!exists(project.getId())) throw new IllegalArgumentException("projectID non presente");
+        checkProject(project);
         return projectRepository.save(project);
     }
 
@@ -106,19 +108,26 @@ public class SimpleProjectsManager implements ProjectsManager{
 
     @Override
     public boolean submit(String projectId, Role role) throws EntityNotFoundException {
-        //TODO 09/02/2021 controllare il contenuto di role
         if(projectId.isBlank()) throw new IllegalArgumentException("Il campo 'projectId' è vuoto");
-        if(Objects.isNull(role)) throw new IllegalArgumentException("Il campo 'role' è nullo");
+        checkRole(role);
         Project project = projectRepository.findById(projectId).orElseThrow(()->
                 new EntityNotFoundException("Nessun progetto trovato con l'id: "+projectId));
         if(!project.submit(role)) return false;
         return projectRepository.save(project).getCandidates().contains(role);
     }
 
+    private void checkRole(Role role){
+        Objects.requireNonNull(role, "Il campo role è nullo");
+        if(role.getUserMail().isBlank()) throw new IllegalStateException("Il campo userMail di role è vuoto");
+        Objects.requireNonNull(role.getSkill(), "Il campo Skill di role è nullo");
+        if(role.getSkill().getName().isBlank()) throw new IllegalStateException("Il campo name del campo skill  di role è vuoto");
+        if(role.getSkill().getLevel() > 10 || role.getSkill().getLevel() <= 0) throw new IllegalStateException("Il campo name del campo skill  di role è vuoto");
+    }
+
     @Override
     public boolean acceptCandidate(String projectId, Role userRole) throws EntityNotFoundException {
         if(projectId.isBlank()) throw new IllegalArgumentException("Il campo 'projectId' è vuoto");
-        if(Objects.isNull(userRole)) throw new IllegalArgumentException("Il campo 'userRole' è nullo");
+        checkRole(userRole);
         Project project = projectRepository.findById(projectId).orElseThrow(()->
                 new EntityNotFoundException("Nessun progetto trovato con l'id: "+projectId));
         project.acceptCandidate(userRole);
@@ -128,7 +137,7 @@ public class SimpleProjectsManager implements ProjectsManager{
     @Override
     public boolean rejectCandidate(String projectId, Role userRole) throws EntityNotFoundException {
         if(projectId.isBlank()) throw new IllegalArgumentException("Il campo 'projectId' è vuoto");
-        if(Objects.isNull(userRole)) throw new IllegalArgumentException("Il campo 'userRole' è nullo");
+        checkRole(userRole);
         Project project = projectRepository.findById(projectId).orElseThrow(()->
                 new EntityNotFoundException("Nessun progetto trovato con l'id: "+projectId));
         project.rejectCandidate(userRole);
@@ -140,7 +149,6 @@ public class SimpleProjectsManager implements ProjectsManager{
         if(userMail.isBlank()) throw new IllegalArgumentException("Il campo 'userMail' è vuoto");
         userRepository.findById(userMail).orElseThrow(()->
                 new EntityNotFoundException("Nessun utente trovato con l'email: "+userMail));
-        //TODO 09/02/2021 da rifare se possibile con gli stream
         List<Role> userSubmissions = new ArrayList<>();
         for(Project p: projectRepository.findAll())
             for(Role r: p.getCandidates())
